refine test-infra
