+++ test-infra/tests/cdc/cdc.go
@@ -173,3 +173,4 @@ func workload(sourceAddr string, targetAddr string, concurrency int, dbName stri
 func genDbDsn(addr string, dbName string) string {
 	return fmt.Sprintf("root:@tcp(%s)/%s", addr, dbName)
 }
+
